Custom actions in index_as_sortable

+++ lib/active_admin/views/index_as_sortable.rb
@@ -39,8 +39,10 @@ module ActiveAdmin
       end
 
       # Adds links to View, Edit and Delete
-      def default_actions
-        @default_actions = true
+      def actions(options = {}, &block)
+        options = { :defaults => true }.merge(options)
+        @default_actions = options[:defaults]
+        @other_actions = block
       end
 
 
@@ -84,7 +86,7 @@ module ActiveAdmin
               call_method_or_proc_on(item, @label)
             end
             div :class => "cell right" do
-              build_default_actions(item) if @default_actions
+              build_actions(item)
             end
           end
 
@@ -96,17 +98,20 @@ module ActiveAdmin
         end
       end
 
-      def build_default_actions(resource)
+      def build_actions(resource)
         links = ''.html_safe
-        if controller.action_methods.include?('show')
-          links << link_to(I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link")
+        if @default_actions
+          if controller.action_methods.include?('show')
+            links << link_to(I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link")
+          end
+          if controller.action_methods.include?('edit')
+            links << link_to(I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link")
+          end
+          if controller.action_methods.include?('destroy')
+            links << link_to(I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')}, :class => "member_link delete_link")
+          end
         end
-        if controller.action_methods.include?('edit')
-          links << link_to(I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link")
-        end
-        if controller.action_methods.include?('destroy')
-          links << link_to(I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')}, :class => "member_link delete_link")
-        end
+        links << instance_exec(resource, &@other_actions) if @other_actions
         links
       end
